there was far more stuff in here than i planned Attack stuff, file reorder
