+++ Assets/Scripts/Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 087796d230a0ed144a396efbdba0fb09
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
