+++ Assets/Scripts/Animation/IAnimationComplete.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using Assets.Scripts.Animation;
+using UnityEngine;
+
+namespace Animation
+{
+    public interface IAnimationComplete
+    {
+        public string ValueName
+        {
+            set => ValueHash = Animator.StringToHash(value);
+        }
+
+        public int ValueHash { get; set; }
+        public EventHandler<AnimatorStateEventArgs> OnComplete { get; set; }
+
+    }
+}
