+++ Assets/Scripts/Animation/IAnimationComplete.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: faadb51a23d89fc449b076d224945c7f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
