+++ Assets/Scripts/Player/AttackEndBehavior.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e6bdb954a50ad0d4db12c53d01f643c6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
