+++ Assets/Scripts/Player/InputControls/PlayerInputControls.cs
@@ -229,7 +229,7 @@ namespace RainbowJam.Controls
                     ""path"": ""<Keyboard>/l"",
                     ""interactions"": """",
                     ""processors"": """",
-                    ""groups"": """",
+                    ""groups"": ""Keyboard&Mouse"",
                     ""action"": ""Fire"",
                     ""isComposite"": false,
                     ""isPartOfComposite"": false
