+++ Assets/Sprites/Placeholders/PlayerParts/PlayerAttack.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 03af88504f2884b499f2831aec928e1b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
