+++ Assets/Sprites/Placeholders/Trowel.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:b94430055ce4be76306675bd7cf461fcd92c747149a3b1062ec4e5720c689c79
+size 7720
